export `Message` and dependants (CloseFrame, CloseCode) by default
