+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name          = "mews"
-version       = "0.1.1"
+version       = "0.1.2"
 edition       = "2021"
 authors       = ["kanarus <kanarus786@gmail.com>"]
 documentation = "https://docs.rs/mews"
